refactor(compiler-plugin): inline registry/ subpackage + fix README compat path
The `registry/` subpackage held only two classes (`PreviewLabFirExtensionRegistrar`
and `PreviewLabIrGenerationExtension`), each used by exactly one site
(`ComposePreviewLabCompilerPluginRegistrar`). The extra layer separated the
abstract-base ownership but added no concrete organizational value — both
classes already belong to the plugin-level "register everything" concern.

Move both files up to the `compiler` root package next to
`ComposePreviewLabCompilerPluginRegistrar` so the registration trio is
collocated, drop the now-empty `registry/` directory.

Also refresh the README directory tree:

- Reflect the registry/ inlining.
- Correct the compat module paths: they are listed as `compiler-plugin-compat-k*`
  but in the repository they live under `compiler-plugin/compat-k*/`
  (subprojects `:compiler-plugin:compat-k*`). Update the tree, the compat
  layer prose, and the `under registry/` reference text.

No behaviour change. Addresses PR #200 review threads:
- PRRT_kwDOOftTdM6A8JKB (registry/ inline)
- PRRT_kwDOOftTdM6A8Ipy (README compat module path)
